final pull

+++ destroy-vm.py
@@ -19,10 +19,19 @@ sys.path.append(os.path.dirname(os.path.realpath(__file__)) + '/lib')
 import api_ovirt
 import api_foreman
 import api_zookeeper
+import api_vmware
 
 vm_config = None
 zk_base_path = '/puppet'
 from config import Config
+import simplecrypt
+import base64
+#disable warnings urllib
+import warnings
+warnings.simplefilter('ignore')
+#disable warnings Foreman API
+import logging
+logging.disable(logging.ERROR)
 
 def deleteZookeeperPath(path, recursive=False):
     zk = kazoo.client.KazooClient(hosts=vm_config.zookeeper_address + ':' + vm_config.zookeeper_port)
@@ -42,32 +51,42 @@ def getZookeeperValue(path):
 def destroyVMs():
     for vm in vm_config.vm_list:
         vm_info = vm_config.vm_list[vm]
-
         print " - Connect to hypervisor"
-        ovirt_conn = api_ovirt.connectToHost(vm_info["hypervisor"], vm_info["hypervisor_user"], vm_info['hypervisor_password'])
+        if vm_info['hypervisor_type'].lower() in ['ovirt', 'rhev']:
+            hypervisor_conn = api_ovirt.connectToHost(vm_info["hypervisor"], vm_info["hypervisor_user"], simplecrypt.decrypt(vm_config.salt, base64.b64decode(vm_info['hypervisor_password'])))
+        else:
+            hypervisor_conn = api_vmware.connectToHost(vm_info["hypervisor"], vm_info["hypervisor_user"], simplecrypt.decrypt(vm_config.salt, base64.b64decode(vm_info['hypervisor_password'])))
+        #determine how to name VM
+        if vm_config.use_fqdn_as_name == 0:
+            vm_name = vm
+        else:
+            vm_name = vm_info['vm_fqdn']
 
-        print "*" * sum((12, len(vm_info['vm_fqdn'])))
-        print "***** " + vm_info['vm_fqdn'] + " *****"
-        print "*" * sum((12, len(vm_info['vm_fqdn'])))
+        print "*" * sum((12, len(vm_name)))
+        print "***** " + vm_name + " *****"
+        print "*" * sum((12, len(vm_name)))
 
-        result = api_ovirt.destroyGuest(ovirt_conn, vm_info['vm_fqdn'])
-        if result != "Succesfully removed guest: " + vm_info['vm_fqdn']:
-            print result
-            print "Finished unsuccesfully, aborting"
-            ovirt_conn.disconnect()
-            sys.exit(99)
+        if vm_info['hypervisor_type'].lower() in ['ovirt', 'rhev']:
+            result = api_ovirt.destroyGuest(hypervisor_conn, vm_name)
+        else:
+            result = api_vmware.destroyGuest(hypervisor_conn, vm_name)
+#        if result != "Succesfully removed guest: " + vm_name:
+#            print result
+#            print "Finished unsuccesfully, aborting"
+#            hypervisor_conn.disconnect()
+#            sys.exit(99)
         print " -", result
 
         print " - Connect to Foreman"
-        foreman_conn = api_foreman.connectToHost(vm_info["foreman"], vm_info["foreman_user"], vm_info['foreman_password'])
+        foreman_conn = api_foreman.connectToHost(vm_info["foreman"], vm_info["foreman_user"], simplecrypt.decrypt(vm_config.salt, base64.b64decode(vm_info['foreman_password'])))
         result = api_foreman.destroyGuest(foreman_conn, vm_info['vm_fqdn'])
         if result != "Succesfully removed guest: " + vm_info['vm_fqdn']:
-            print result
+            print "   - ", result
             print "Finished unsuccesfully, aborting"
             sys.exit(99)
-        print " -", result
+        print "   - ", result
         print " - Connect to Puppetmaster"
-        puppet_conn = puppet.Puppet(host='puppetmaster01.core.cmc.lan', port=8140, key_file='./ssl/api-key.pem', cert_file='./ssl/api-cert.pem')
+        puppet_conn = puppet.Puppet(host=vm_config.puppetmaster_address, port=vm_config.puppetmaster_port, key_file='./ssl/api-key.pem', cert_file='./ssl/api-cert.pem')
         print "   - Clear certificate of %s" % vm_info['vm_fqdn']
         try:
             puppet_conn.certificate_clean(vm_info['vm_fqdn'])
@@ -78,15 +97,16 @@ def destroyVMs():
         print " - Connect to zookeeper"
         zookeeper_conn = api_zookeeper.connectToHost(vm_config.zookeeper_address, vm_config.zookeeper_port)
         print " - Zookeeper records"
-        print "   - Get ossec server IP"
-        ossecserver, nodeStats = api_zookeeper.getValue(zookeeper_conn, zk_base_path + '/' + vm_info['puppet_environment'] + '/defaults/profile::ossec::client::ossec_server')
-        print "   - Delete zookeeper ossec auth"
-        zk_path = zk_base_path + '/production/nodes/' + ossecserver + '/client-keys/' + vm_info['vm_fqdn']
-        result = api_zookeeper.deleteValue(zookeeper_conn, zk_path, recursive=True)
-        if result != "Succesfully deleted path '%s'" % zk_path:
-            print result
-            print "Finished unsuccesfully, aborting"
-            sys.exit(99)
+        if vm_info['ossec_in_env'] == 1:
+            print "   - Get ossec server IP"
+            ossecserver, nodeStats = api_zookeeper.getValue(zookeeper_conn, zk_base_path + '/' + vm_info['puppet_environment'] + '/defaults/core::profile::ossec::client::ossec_server')
+            print "   - Delete zookeeper ossec auth"
+            zk_path = zk_base_path + '/production/nodes/' + ossecserver + '/client-keys/' + vm_info['vm_fqdn']
+            result = api_zookeeper.deleteValue(zookeeper_conn, zk_path, recursive=True)
+            if result != "Succesfully deleted path '%s'" % zk_path:
+                print result
+                print "Finished unsuccesfully, aborting"
+                sys.exit(99)
         print "   - Delete zookeeper puppet node"
         zk_path = zk_base_path + '/' + vm_info['puppet_environment'] + '/nodes/' + vm_info['vm_fqdn']
         result = api_zookeeper.deleteValue(zookeeper_conn, zk_path, recursive=True)
@@ -94,8 +114,25 @@ def destroyVMs():
             print result
             print "Finished unsuccesfully, aborting"
             sys.exit(99)
+    if vm_config.vm_type.lower() == 'oracle-rac':
+        #delete shared disks from first vm
+        for vm in sorted(vm_config.vm_list.keys()):
+            vm_info = vm_config.vm_list[vm]
+            #determine how to name VM
+            if vm_config.use_fqdn_as_name == 0:
+                vm_name = vm
+            else:
+                vm_name = vm_info['vm_fqdn']
+
+            disk_counter = 1
+            if vm_info['hypervisor_type'].lower() in ['ovirt', 'rhev']:
+                for disk in vm_config.shared_disks:
+                    api_ovirt.deleteDisk(hypervisor_conn, vm_name+'_racdisk'+str(disk_counter).zfill(2))
+                    disk_counter += 1
+            break
+
     print " - Disconnect from hypervisor"
-    ovirt_conn.disconnect()
+    hypervisor_conn.disconnect()
     print " - Disconnect from zookeeper"
     api_zookeeper.disconnect(zookeeper_conn)
